% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/generate_snp_data.R
\name{generate_snp_data}
\alias{generate_snp_data}
\title{SNP Data Generator}
\usage{
generate_snp_data(map_file, ped_file)
}
\arguments{
\item{map_file}{The name of the .map file. Should be tab delimited with no header.}

\item{ped_file}{The name of the .ped file. Should be space delimited with no header.}
}
\description{
Takes in a plink map file and a plink ped file and creates a dataframe with headers
}
\examples{
generate_snp_data("HapMap_Mito_Seq_QC_2.map","HapMap_Mito_Seq_QC_2.ped")
}
\keyword{map}
\keyword{ped}
\keyword{snp}

