% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getClassifications.R
\name{getClassifications}
\alias{getClassifications}
\title{getClassifications output generator}
\usage{
getClassifications(source_df)
}
\arguments{
\item{source_df}{The snp_data generated dataframe}
}
\description{
Takes in a dataframe generated by "generate_snp_data" and returns each subject classified with full classification paths
}
\keyword{classification}
\keyword{haplogroup}
\keyword{list}
\keyword{path}

