% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getFinalPathList.R
\name{getFinalPathList}
\alias{getFinalPathList}
\title{getFinalPathList internal function}
\usage{
getFinalPathList(list)
}
\arguments{
\item{df}{The FASTMAP dataframe row}
}
\description{
Takes in a nested list of prettified paths and returns those of the greatest length
}
\keyword{list}
\keyword{path}

