% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getGroupFromPath.R
\name{getGroupFromPath}
\alias{getGroupFromPath}
\title{getFinalPathList internal function}
\usage{
getGroupFromPath(string)
}
\arguments{
\item{string}{The path in question}
}
\description{
Takes in a list of paths and returns the final classification of that path
}
\keyword{classification}
\keyword{haplogroup}
\keyword{list}
\keyword{path}

