% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getPathList.R
\name{getPathList}
\alias{getPathList}
\title{getPathList internal function}
\usage{
getPathList(df)
}
\arguments{
\item{df}{The FASTMAP dataframe row}
}
\description{
Takes in a nested list of paths in dataframe format and returns a top-level path assignment
}
\keyword{list}
\keyword{path}
\keyword{prettify}

