% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\docType{data}
\name{node_b4bde}
\alias{node_b4bde}
\title{A node node_b4bde.}
\format{An object of class \code{node} of length 1.}
\usage{
node_b4bde
}
\description{
An instance of the node class.
}
\keyword{datasets}
