% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\docType{data}
\name{node_u8b_alt}
\alias{node_u8b_alt}
\title{A node node_u8b_alt.}
\format{An object of class \code{node} of length 1.}
\usage{
node_u8b_alt
}
\description{
An instance of the node class.
}
\keyword{datasets}
