% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/numChildren.R
\name{numChildren}
\alias{numChildren}
\title{NumChildren}
\usage{
numChildren(node_object)
}
\arguments{
\item{node_object}{The FASTMAP node in question}
}
\description{
Internal function. Takes in a node object and returns the total number of that node's children
}
\keyword{children}
\keyword{node}
\keyword{snp}

