% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numReqs.R
\name{numReqs}
\alias{numReqs}
\title{NumReqs}
\usage{
numReqs(node_object)
}
\arguments{
\item{node_object}{The FASTMAP node in question}
}
\description{
Internal function. Takes in a node object and returns the number of SNPs that it requires for validation
}
\examples{
#internal function, not meant to be called externally
number_of_required_snps <- numReqs(HiMC::root)
}
\keyword{node}
\keyword{snp}
