% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validData.R
\name{validData}
\alias{validData}
\title{validData boolean check}
\usage{
validData(df, node)
}
\arguments{
\item{df}{The FASTMAP dataframe row}

\item{node}{The node in question}
}
\description{
Internal function. Takes in a dataframe and a node and returns true if the dataframe row has the required SNPs for the node
}
\examples{
#internal function, not meant to be called externally
df_row_meets_criteria_for_node_requirements <- validData(data.frame(),HiMC::root)
}
\keyword{data}
\keyword{df}
\keyword{node}
\keyword{snp}
\keyword{valid}
