% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathList.R
\name{getPathList}
\alias{getPathList}
\title{getPathList internal function}
\usage{
getPathList(df)
}
\arguments{
\item{df}{The FASTMAP dataframe row}
}
\description{
Takes in a nested list of paths in dataframe format and returns a top-level path assignment
}
\examples{
#internal function, not meant to be called externally
paths <- getPathList(data.frame())
}
\keyword{list}
\keyword{path}
\keyword{prettify}
